Add createSlackMessage function and refactor sendMessage to use it
