+++ src/messages/send.ts
@@ -9,6 +9,7 @@ import { Relay, useWebSocketImplementation } from 'nostr-tools/relay';
 import pino from 'pino';
 import WebSocket from 'ws';
 import env from '../env';
+import { createSlackMessage } from './create';
 
 useWebSocketImplementation(WebSocket);
 
@@ -104,41 +105,7 @@ export default async function sendMessage(
 
   // Post to Slack
   if (env.SLACK_BOT_TOKEN !== undefined && env.SLACK_CHANNEL_ID !== undefined) {
-    const lines = text.split('\n').filter((line) => line.trim() !== '');
-
-    const maxLine = lines.find((line) =>
-      line.startsWith('Maximum Seismic Intensity'),
-    );
-    const max = maxLine
-      ? Number.parseInt(maxLine.replace('Maximum Seismic Intensity ', ''), 10)
-      : null;
-
-    const area = new Map<string, string>();
-
-    for (const line of lines.slice(2)) {
-      const match = line.match(/\[Seismic Intensity (\d)\] (.+)/);
-      if (match) {
-        const intensity = match[1];
-        const regions = match[2];
-        area.set(intensity, regions);
-      }
-    }
-
-    const attachments = [
-      {
-        fallback: `${lines[0]}: Maximum Seismic Intensity ${max}`,
-        color: '#228BFF',
-        title: lines[0],
-        text: `Maximum Seismic Intensity ${max}`,
-        fields: [
-          ...Array.from(area.entries()).map(([intensity, regions]) => ({
-            title: `Seismic Intensity ${intensity}`,
-            value: regions,
-            short: true,
-          })),
-        ],
-      },
-    ];
+    const attachments = createSlackMessage(text);
 
     try {
       await slackClient.chat.postMessage({
